Files and Flags.
I've changed it to use the cmdline library.
Now it doesn't require a file called Schedule.txt, and only output to a file called Schedule.tex.

It accepts command line flags for input and output files, and it doesn't find any, it assumes stdin for input and stdout for output.

This is already a vast improvement.