Editable Title
What I've done is make any line that isn't a comment, a class spec, or a time spec into a title.

This means that if I were to have a line that doesn't quite fit the spec, it would be made into a title.
This should be okay. It was just ignored before.

It's one of those things. Maybe someone will notice this at some point, but until then it will just seem to work.

Also, I've made the date spec a little more flexible, and a leading space is acceptable too, not just a tab.

I've moved all output code from the Setup function, where it didn't belong, to the PrintChart function, where it does belong.
It had to be done because Setup was called before the file was read, and that was where the title was being output.
This will also make the next step nicer, where I make multiple output formats.