Better Seperation and Commenting.
I've separated the Output from Input.
It's actually pretty cool.

Now, the Setup function doesn't need any file pointers, the ReadFile only needs an In file, and PrintChart only needs Out.
This is how it should have been from the start, but wasn't.

I've had to move some code from ReadFile to PrintChart, along with some new things.
The ReadFile was writting macros to the file as it found courses and addTime was just writting the macros into the Array.
This has been stopped now, as I want to move away from LaTeX.

I think all Tex specific things are in the PrintChart now, so I can move forward with HTML and others.

Also, I've added some comments, especially in the TimeIndex function, which was a little dense and not explained.
It took me a second to figure out what was going on there.